% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_card.R
\name{add_card_border}
\alias{add_card_border}
\title{Helper to add a border to each card plot in a list}
\usage{
add_card_border(
  plots,
  card = NULL,
  inset = grid::unit(c(5, 5), "mm"),
  fill = NA,
  color = "white",
  linetype = "dashed",
  linewidth = 1
)
}
\arguments{
\item{inset}{Numeric distance to inset an input page.}

\item{fill, color, linetype, linewidth}{Fixed aesthetics passsed to
\code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}}
}
\description{
\code{\link[=add_card_border]{add_card_border()}} is a helper function used internally by \code{\link[=plot_cards]{plot_cards()}}
and can be used to further modify output plots.
}
\keyword{card}
\keyword{internal}
