% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_card.R
\name{add_card_number}
\alias{add_card_number}
\title{Helper to add a number to each card plot in a list}
\usage{
add_card_number(
  plots,
  card = NULL,
  n = 1,
  color = "white",
  size = 5,
  family = NULL,
  sequence = TRUE
)
}
\description{
\code{\link[=add_card_number]{add_card_number()}} is a helper function used internally by \code{\link[=plot_cards]{plot_cards()}} to
apply a number to each card in a list of plots.
}
\keyword{card}
\keyword{internal}
