% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_card.R
\name{add_card_text}
\alias{add_card_text}
\title{Helper to add text to each card plot in a list}
\usage{
add_card_text(
  plots,
  card = NULL,
  text = NULL,
  color = "white",
  size = 5,
  family = NULL,
  nudge_x = 0,
  nudge_y = 0.5
)
}
\description{
\code{\link[=add_card_text]{add_card_text()}} is a helper function used internally by \code{\link[=plot_cards]{plot_cards()}}.
}
\keyword{card}
\keyword{internal}
