% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{card_sizes}
\alias{card_sizes}
\title{Standard card sizes}
\format{
A data frame with 5 rows and 6 variables:
\describe{
\item{\code{name}}{Name of card}
\item{\code{units}}{Units ("in" or "mm") for dimensions}
\item{\code{width}}{Width in units}
\item{\code{height}}{Height in units}
\item{\code{orientation}}{Portrait (width less than height), landscape, or
square}
}
}
\usage{
card_sizes
}
\description{
Reference table of common playing card sizes for \code{\link[=get_card]{get_card()}}. Data is a
subset of \code{paper_sizes} which is also included with \code{{sfext}} package.
}
\keyword{datasets}
