% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards-package.R
\docType{package}
\name{cards-package}
\alias{cards}
\alias{cards-package}
\title{cards: Plot Playing Cards}
\description{
Make simple playing cards with ggplot2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/cards}
  \item \url{https://elipousson.github.io/cards/}
  \item Report bugs at \url{https://github.com/elipousson/cards/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
