% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dist_scale.R
\name{convert_dist_scale}
\alias{convert_dist_scale}
\title{Convert distance from scale to actual units}
\usage{
convert_dist_scale(
  dist = NULL,
  scale = NULL,
  standard = NULL,
  series = NULL,
  scale_unit = "in",
  scale_factor = NULL,
  actual_unit = NULL,
  dpi = 120,
  paper = NULL,
  orientation = NULL,
  ...
)
}
\arguments{
\item{dist}{distance to convert. If paper is provided, dist is optional and
paper width and height are used as dist.}

\item{scale}{Scale name from \code{standard_scales[["scale"]]}.}

\item{standard}{Scale standard. Options include "USGS", "Engineering", or
"Architectural".}

\item{series}{Map series from \code{standard_scales[["series"]]}. Series is only
available for USGS scales.}

\item{scale_unit}{"mm" (converted to cm by dividing by 10), "cm", "px"
(converted to inches by dividing by dpi), or "in".}

\item{scale_factor}{factor for converting from scale_unit to actual_unit,
e.g. if 1" = 1', the scale factor is 12. optional if scale if provided;
defaults to \code{NULL}.}

\item{actual_unit}{any unit supported by \code{\link[=convert_dist_units]{convert_dist_units()}}}

\item{dpi}{dots per square inch (used as conversion factor for "px" to "in")}

\item{paper}{Name of paper passed to \code{\link[=get_paper]{get_paper()}}}

\item{orientation}{Page orientation, Default: \code{NULL}. Supported options are
"portrait", "landscape", or "square".}

\item{...}{
  Arguments passed on to \code{\link{get_paper}}
  \describe{
    \item{\code{name}}{Page name, e.g. "letter", not case sensitive, Default: \code{NULL}}
    \item{\code{width}}{Page width in "in", "px" or "mm" units. Default: \code{NULL}}
    \item{\code{height}}{Page height in "in", "px" or "mm" units. Default: \code{NULL}}
    \item{\code{reorient}}{If \code{TRUE} and orientation is not \code{NULL}, flip width and height
dimensions for any pages that do not match the provided orientation. Set
\code{reorient = FALSE} to filter pages by orientation.}
    \item{\code{type}}{Page type, Options include "paper", "social", "postcard",
"print", "card", or "screen". Default: \code{NULL}}
    \item{\code{ignore.case}}{If \code{FALSE}, filtering for page and type are case
sensitive. Defaults to \code{TRUE}.}
    \item{\code{page}}{Used by \code{\link[=get_page_dims]{get_page_dims()}}, page is either a character vector
passed to the name parameter of \code{\link[=get_page_size]{get_page_size()}}, a data.frame with column
names matching the cols parameter, or a length 2 numeric vector with the
page width and height.}
    \item{\code{cols}}{Length 2 character vector with column names for page dimensions.
Defaults to c("width", "height").}
    \item{\code{arg,call}}{Passed to \code{\link[cli:cli_abort]{cli_abort()}} to improve internal error messages.}
    \item{\code{units}}{Units to convert page dimensions to using \code{\link[=convert_unit_type]{convert_unit_type()}}.}
    \item{\code{valueOnly}}{Passed to valueOnly parameter of \code{\link[grid:convertUnit]{grid::convertUnit()}}.
Defaults to \code{FALSE}.}
  }}
}
\value{
\itemize{
\item If paper is not provided, return a vector of dist values converted from
scale_unit to actual_unit based on scale_factor or information from
\link{standard_scales} data.
\item If paper is provided, return a data.frame with converted distances appends
as columns named actual_width and actual_height.
}
}
\description{
This function converts scale distances to actual units based on named
\link{standard_scales}.
}
\seealso{
Other dist:
\code{\link[=convert_dist_units]{convert_dist_units()}},
\code{\link[=is_dist_units]{is_dist_units()}}
}
\concept{dist}
