% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page.R
\name{get_page}
\alias{get_page}
\alias{get_paper}
\alias{get_card}
\title{Get a paper or card size based on name, dimensions, orientation, or type}
\usage{
get_page(
  name = NULL,
  width = NULL,
  height = NULL,
  orientation = NULL,
  type = NULL
)

get_paper(name = NULL, width = NULL, height = NULL, orientation = NULL)

get_card(name = NULL, width = NULL, height = NULL, orientation = NULL)
}
\arguments{
\item{name}{Page name, e.g. "letter", not case sensitive, Default: NULL}

\item{width}{Page width in "in", "px" or "mm" units. Default: NULL}

\item{height}{Page height in "in", "px" or "mm" units. Default: NULL}

\item{orientation}{Page orientation, Default: NULL}

\item{type}{Page type, Options include "paper", "social", "postcard",
"print", "card", or "screen". Default: NULL}
}
\value{
A data.frame with page, paper, or card name and dimensions.
}
\description{
Get a paper or card size based on name, dimensions, orientation, or type
}
