% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_social_size.R
\name{get_social_size}
\alias{get_social_size}
\title{Get social media image size to match platform and format}
\usage{
get_social_size(
  name = NULL,
  platform = NULL,
  format = NULL,
  orientation = NULL
)
}
\arguments{
\item{name}{Image size name, Default: \code{NULL}}

\item{platform}{Social media platform, "Instagram", "Facebook", or "Twitter",
Default: \code{NULL}}

\item{format}{Image format, "post", "story", or "cover", Default: \code{NULL}}

\item{orientation}{Image orientation, Default: \code{NULL}.}
}
\description{
See \code{paper_sizes[paper_sizes$type == "social",]$name} for support image
options.
}
\examples{
get_social_size("Instagram post")

get_social_size(platform = "Twitter")

get_social_size(format = "cover")

}
\seealso{
\code{\link[=get_page_size]{get_page_size()}}
}
