% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_units}
\alias{grid_units}
\title{Grid units (vector)}
\format{
A character vector with 34 abbreviated, singular, and plural unit
names.
}
\usage{
grid_units
}
\description{
A vector of units supported by \code{\link[grid:unit]{grid::unit()}}
}
\keyword{datasets}
