% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset_page.R
\name{inset_page}
\alias{inset_page}
\title{Apply numeric or unit class inset to width/height dimensions of page}
\usage{
inset_page(
  page,
  inset = 0.1,
  units = "in",
  pct_inset = NULL,
  cols = c("width", "height")
)
}
\arguments{
\item{inset}{Numeric distance to inset an input page.}

\item{units}{A character vector specifying the units for the
    corresponding numeric values.}

\item{pct_inset}{Percent inset based on existing page dimensions.}
}
\description{
\code{\link[=inset_page]{inset_page()}} applies a set inset distance value or percent distance to
modify the dimensions of the input page list or data frame.
}
\keyword{internal}
