% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_cards.R
\name{layout_cards}
\alias{layout_cards}
\title{Use patchwork to layout cards on a paper}
\usage{
layout_cards(
  card = NULL,
  paper = NULL,
  width = NULL,
  height = NULL,
  orientation = "landscape",
  byrow = FALSE,
  ...
)
}
\arguments{
\item{card}{Card name, a data.frame with width and height columns, or a list
of ggplot2 objects with card plots. Default: \code{NULL}}

\item{paper}{Paper name or a data.frame with width and height columns.
Optional if width and height are both provided, Default: \code{NULL}}

\item{width, height}{Paper width and height, Default: \code{NULL}}

\item{orientation}{Paper orientation, Optional if width and height are both
provided, Default: 'landscape'}

\item{byrow}{Analogous to \code{byrow} in \link[base:matrix]{matrix()}. If \code{FALSE} the
plots will be filled in in column-major order}

\item{...}{
  Arguments passed on to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots}}
  \describe{
    \item{\code{widths,heights}}{The relative widths and heights of each column and row
in the grid. Will get repeated to match the dimensions of the grid.}
    \item{\code{guides}}{A string specifying how guides should be treated in the layout.
\code{'collect'} will collect guides below to the given nesting level, removing
duplicates. \code{'keep'} will stop collection at this level and let guides be
placed alongside their plot. \code{auto} will allow guides to be collected if a
upper level tries, but place them alongside the plot if not.  If you modify
default guide "position" with \link[ggplot2:theme]{theme(legend.position=...)}
while also collecting guides you must apply that change to the overall
patchwork (see example).}
    \item{\code{tag_level}}{A string (\code{'keep'} or \code{'new'}) to indicate how
auto-tagging should behave. See \code{\link[patchwork:plot_annotation]{plot_annotation()}}.}
    \item{\code{design}}{Specification of the location of areas in the layout. Can either
be specified as a text string or by concatenating calls to \code{\link[patchwork:area]{area()}} together.
See the examples for further information on use.}
  }}
}
\value{
A \code{patchwork} object
}
\description{
Use patchwork to layout cards on a paper
}
\examples{
layout_cards(
  card = plot_cards("Poker", 6),
  paper = "letter"
)
}
\seealso{
\code{\link[ggplot2:ggplot_build]{ggplot2::ggplot_build()}}
\code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}, \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}
}
