% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{page_extras}
\alias{page_extras}
\title{Extra reference data for page layouts}
\format{
A length 1 list.
}
\usage{
page_extras
}
\description{
A named list of additional reference data that currently includes only one
data.frame: a reference table of margin sizes in "in" and "cm".
}
\keyword{datasets}
