% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papersize-package.R
\docType{package}
\name{papersize-package}
\alias{papersize}
\alias{papersize-package}
\title{papersize: Sizing Plots and Files for Paper}
\description{
A set of convenience functions extending grid, ggplot2, and patchwork to help you size plots and files for printing to paper, postcards, playing cards, and other physical media.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/papersize}
  \item \url{https://elipousson.github.io/papersize/}
  \item Report bugs at \url{https://github.com/elipousson/papersize/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]
}

}
\keyword{internal}
