% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_card.R
\name{plot_cards}
\alias{plot_cards}
\title{Use ggplot to plot for one or more cards}
\usage{
plot_cards(
  card,
  n = 1,
  orientation = "portrait",
  number = FALSE,
  color = "white",
  size = 5,
  family = NULL,
  fill = "gray20",
  border = FALSE,
  inset = unit(c(5, 5), "mm"),
  linetype = "dashed",
  linewidth = 1,
  text = NULL,
  center = c(0, 0)
)
}
\arguments{
\item{card}{Card name or data.frame with width and height columns.}

\item{n}{Number of cards to plot, Default: 1}

\item{orientation}{Card orientation, Default: 'portrait'}

\item{number}{If \code{TRUE}, add a number to each card, Default: \code{FALSE}}

\item{color}{Color for number, text, and border, Default: 'white'}

\item{size}{Font size for number and text, Default: 5}

\item{family}{Font family for number and text, Default: 'Georgia'}

\item{fill}{Length 1 or 2 character vector with color name. If length 2, the
first value is assumed to be the card fill and the second value is assumed
to be the inset border fill. Default: 'gray20'}

\item{border}{If TRUE, add a border to the card. Default: FALSE}

\item{inset}{Unit or numeric vector with inset distance for card border,
Default: unit(c(5, 5), "mm"). If inset is a numeric vector, it is expected
to be a percent relative to the card width and height.}

\item{linetype}{linetype for card border, Default: 'dashed'}

\item{linewidth}{linewidth for card border, Default: 2}

\item{text}{Character vector with card text, Default: NULL}

\item{center}{Position of card center, Default: c(0, 0)}
}
\value{
A list of plot objects where each item on the list is a card.
}
\description{
Make a plot of cards.
}
\examples{
plot_cards("Tarot", n = 2, number = TRUE)[[2]]

plot_cards("Poker", n = 1, number = TRUE, text = "♡️")
}
