% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_to_page.R
\name{print_to_page}
\alias{print_to_page}
\alias{print_to_page_layout}
\title{Explicitly draw plot using dimensions from page data.frame or list}
\usage{
print_to_page(plot, page, newpage = TRUE, vp = NULL, ...)

print_to_page_layout(
  plot,
  page,
  row_position = 1,
  col_position = 1,
  row_height = NULL,
  col_width = NULL,
  nrow = 1,
  ncol = 1,
  layout = NULL,
  parent = NULL,
  children = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{plot}{Plot to display}

\item{page}{A page data.frame from \code{\link[=get_page_size]{get_page_size()}} or \code{\link[=make_page_size]{make_page_size()}}.}

\item{newpage}{draw new (empty) page first?}

\item{vp}{viewport to draw plot in}

\item{...}{
  Arguments passed on to \code{\link{page_to_viewport}}, \code{\link{page_to_layout}}
  \describe{
    \item{\code{name}}{A character value to uniquely identify the viewport
    once it has been pushed onto the viewport tree.  }
    \item{\code{cols}}{Column names to use for width and height columns. Defaults to
c("width", "height"). Must be length 2 and the first value is always used
as as the width name and the second as the height.}
    \item{\code{margins}}{A numeric list or vector or a margin class object. Defaults to
\code{NULL}. margins are removed from the overall layout width and height.}
    \item{\code{region}}{Optional. An additional page data.frame where the region width
and height are used as the column width and row height.}
    \item{\code{gutter}}{Gutter width/height. Not yet implemented.}
    \item{\code{units}}{Passed to default.units parameter of \code{\link[grid:grid.layout]{grid::grid.layout()}}.}
    \item{\code{widths}}{A numeric vector or unit object
    describing the widths of the columns
    in the layout.}
    \item{\code{heights}}{A numeric vector or unit object
    describing the heights of the rows
    in the layout.}
    \item{\code{respect}}{A logical value or a numeric matrix.
    If a logical, this indicates whether
    row heights and column widths should respect each other.
    If a matrix, non-zero values indicate that the corresponding
    row and column should be respected (see examples below).
  }
    \item{\code{just}}{A string or numeric
    vector specifying how the
    layout should be
    justified if it is not the same size as its parent viewport.
    If there are two values, the first
    value specifies horizontal justification and the second value specifies
    vertical justification.  Possible string values are: \code{"left"},
    \code{"right"}, \code{"centre"}, \code{"center"}, \code{"bottom"},
    and \code{"top"}.  For numeric values, 0 means left alignment
    and 1 means right alignment. NOTE that in this context,
    \code{"left"}, for example, means align the left
    edge of the left-most layout column with the left edge of the
    parent viewport.}
  }}

\item{row_position, col_position}{Row and column position. If nrow is smaller
than row_position or ncol is smaller than col_position, row_position is
used for nrow or col_position is used for ncol instead.}

\item{row_height, col_width}{Row height and column width.}

\item{nrow}{An integer describing the number of rows in the layout.}

\item{ncol}{An integer describing the number of columns in the layout.}

\item{layout}{Passed to \code{\link[=page_to_viewport]{page_to_viewport()}} with layout_position as
layout.pos.row and layout.pos.col if provided. Defaults to \code{NULL} where
layout is defined by page_to_layout using ncol, nrow, page and any
additional parameters passed to \code{...}}

\item{parent}{A grid viewport object.}

\item{children}{A vpList object.}

\item{filename}{File name to create on disk.}
}
\value{
Invisibly returns the original plot using a viewport from
\code{\link[=page_to_viewport]{page_to_viewport()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
if (interactive() && is_installed("ggplot2")) {
  library(ggplot2)

  p <- ggplot(mpg, aes(displ, hwy, colour = class)) +
    geom_point()

  print_to_page(
    p,
    page = get_page_size("Tarot card")
  )
}
}
}
\seealso{
\code{\link[ggplot2:print.ggplot]{ggplot2::print.ggplot()}}
}
