% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_asp.R
\name{as_asp}
\alias{as_asp}
\title{Convert character string, page name, or page data.frame to numeric aspect ratio}
\usage{
as_asp(
  asp = NULL,
  page = NULL,
  orientation = NULL,
  flipped = FALSE,
  sep = ":",
  cols = c("width", "height"),
  ...
)
}
\arguments{
\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
characters separated by a colon (e.g. "1:3"). If numeric, \code{\link[=as_asp]{as_asp()}}
returns the value of asp without modification.}

\item{page}{If character, page is passed as the name parameter to
\code{\link[=get_page_size]{get_page_size()}}.}

\item{orientation}{Page orientation, Default: \code{NULL}. Supported options are
"portrait", "landscape", or "square".}

\item{flipped}{If \code{TRUE}, return aspect ratio of height to width (y / x),
instead of width to height.}

\item{sep}{Separator character to use if asp is a character vector. Defaults
to ":".}

\item{cols}{Length 2 character vector with column names for page dimensions.
Defaults to c("width", "height").}

\item{...}{
  Arguments passed on to \code{\link[=get_page_size]{get_page_size}}
  \describe{
    \item{\code{name}}{Page name, e.g. "letter", not case sensitive, Default: \code{NULL}}
    \item{\code{width}}{Page width in "in", "px" or "mm" units. Default: \code{NULL}}
    \item{\code{height}}{Page height in "in", "px" or "mm" units. Default: \code{NULL}}
    \item{\code{reorient}}{If \code{TRUE} and orientation is not \code{NULL}, flip width and height
dimensions for any pages that do not match the provided orientation. Set
\code{reorient = FALSE} to filter pages by orientation.}
    \item{\code{type}}{Page type, Options include "paper", "social", "postcard",
"print", "card", or "screen". Default: \code{NULL}}
    \item{\code{ignore.case}}{If \code{FALSE}, filtering for page and type are case
sensitive. Defaults to \code{TRUE}.}
    \item{\code{units}}{Units to convert page dimensions to using \code{\link[=convert_unit_type]{convert_unit_type()}}.}
  }}
}
\value{
A numeric vector.
}
\description{
Convert a character vector with an aspect ratio to a numeric aspect ratio or
get the aspect ratio for a given page that has an aspect ratio column.
}
\examples{
as_asp(8.5 / 11)

as_asp("11:17")

as_asp("3/2", sep = "/")

as_asp(page = "letter")

as_asp(page = "letter", orientation = "landscape")

as_asp(page = "letter", orientation = "portrait", flipped = TRUE)

as_asp(page = make_page_size(8.5, 11, "in"), flipped = TRUE)

}
\seealso{
\code{\link[isstatic:as_orientation]{isstatic::as_orientation()}}
}
