% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_asp.R
\name{as_asp}
\alias{as_asp}
\title{Convert character string, page name, or page data.frame to numeric aspect ratio}
\usage{
as_asp(asp = NULL, page = NULL, orientation = NULL, ...)
}
\arguments{
\item{asp}{Aspect ratio of width to height as a numeric value (e.g. 0.33) or
characters separated by a colon (e.g. "1:3"). If numeric, \code{\link[=as_asp]{as_asp()}}
returns the value of asp without modification.}

\item{page}{If character, page is passed as the name parameter to
\code{\link[=get_page_size]{get_page_size()}}.}

\item{orientation}{Page orientation, Default: NULL}

\item{...}{
  Arguments passed on to \code{\link[=get_page_size]{get_page_size}}
  \describe{
    \item{\code{name}}{Page name, e.g. "letter", not case sensitive, Default: NULL}
    \item{\code{width}}{Page width in "in", "px" or "mm" units. Default: NULL}
    \item{\code{height}}{Page height in "in", "px" or "mm" units. Default: NULL}
    \item{\code{reorient}}{If \code{TRUE} and orientation is not NULL, flip width and height
dimensions for any pages that do not match the provided orientation. Set
\code{reorient = FALSE} to filter pages by orientation.}
    \item{\code{type}}{Page type, Options include "paper", "social", "postcard",
"print", "card", or "screen". Default: NULL}
    \item{\code{ignore.case}}{If \code{FALSE}, filtering for page and type are case
sensitive. Defaults to \code{TRUE}.}
    \item{\code{units}}{Units to convert page dimensions to using \code{\link[=convert_unit_type]{convert_unit_type()}}.}
  }}
}
\description{
Convert character string, page name, or page data.frame to numeric aspect ratio
}
\examples{
as_asp(8.5/11)

as_asp("11:17")

as_asp(paper = "letter")

as_asp(paper = "letter", orientation = "landscape")

}
\seealso{
\code{\link[=as_orientation]{as_orientation()}}
}
