% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_page.R
\name{as_page}
\alias{as_page}
\title{Coerce a character or named vector to a page data.frame using get_page_size()
or make_page_size()}
\usage{
as_page(x, ..., cols = c("width", "height"), class = "data.frame")
}
\arguments{
\item{x}{A character vector passed to the name parameter of
\code{\link[=get_page_size]{get_page_size()}}, a named list of vector passed to dims for
\code{\link[=make_page_size]{make_page_size()}}, or a length 2 numeric vector passed as the width and
height parameter to \code{\link[=make_page_size]{make_page_size()}}. If x is a character string that
does not match any of the names in paper_sizes, x is passed as the name
parameter to \code{\link[=make_page_size]{make_page_size()}}}

\item{...}{
  Arguments passed on to \code{\link{make_page_size}}
  \describe{
    \item{\code{width,height}}{Page width and height. Both are required, if asp is
\code{NULL}. Default to \code{NULL}.}
    \item{\code{units}}{Units for width and height. Required unless units is included in
dims. Passed to \code{\link[=as_unit_type]{as_unit_type()}} to validate.}
    \item{\code{asp}}{Aspect ratio. Required if only one of width or height are
provided.}
    \item{\code{name}}{Optional name for paper size. Recommend avoiding duplication with
existing names in \code{paper_sizes}.}
    \item{\code{dims}}{A list or data.frame that can be used to set width, height,
units, and/or asp.}
    \item{\code{orientation}}{Page orientation, Default: \code{NULL}. Supported options are
"portrait", "landscape", or "square". If width and height suggest a
portrait orientation when orientation = "landscape", the dimensions are
reversed so the page dimensions match the provided orientation.}
    \item{\code{valid_units}}{Character vector with name or symbols for valid units.
Defaults to \code{NULL} but any other unit name or symbol, e.g. "px", is
permitted.}
    \item{\code{call}}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}
  }}

\item{cols}{Column names to use for width and height columns. Defaults to
c("width", "height"). Must be length 2 and the first value is always used
as as the width name and the second as the height.}

\item{class}{Class of return object: "data.frame" (default) or "list" (only
supported when input page size is a single row).}
}
\value{
A data.frame or list object with one or more page dimensions.
}
\description{
Coerce a character or named vector to a page data.frame using get_page_size()
or make_page_size()
}
\examples{
as_page(c(8.5, 11), units = "in")

as_page("letter")

as_page(
  list(name = "letter", w = 8.5, h = 11, units = "in"),
  cols = c("w", "h")
)

}
