% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_page.R
\name{as_page}
\alias{as_page}
\title{Coerce a character or named vector to a page data.frame using get_page_size()
or make_page_size()}
\usage{
as_page(x, ..., cols = c("width", "height"), class = "data.frame")
}
\arguments{
\item{x}{A character vector passed to the name parameter of
\code{\link[=get_page_size]{get_page_size()}}, a named list of vector passed to dims for
\code{\link[=make_page_size]{make_page_size()}}, or a length 2 numeric vector passed as the width and
height parameter to \code{\link[=make_page_size]{make_page_size()}}. If x is a character string that
does not match any of the names in paper_sizes, x is passed as the name
parameter to \code{\link[=make_page_size]{make_page_size()}}}

\item{...}{
  Arguments passed on to \code{\link[=make_page_size]{make_page_size}}
  \describe{
    \item{\code{width,height}}{Page width and height. Both are required, if asp is
\code{NULL}. Default to \code{NULL}.}
    \item{\code{units}}{Units for width and height. Required unless units is included in
dims. Passed to \code{\link[=as_unit_type]{as_unit_type()}} to validate.}
    \item{\code{asp}}{Aspect ratio. Required if only one of width or height are
provided.}
    \item{\code{name}}{Optional name for paper size. Recommend avoiding duplication with
existing names in \code{paper_sizes}.}
    \item{\code{cols}}{Column nmaes ot use for width and height columns.}
    \item{\code{dims}}{A list or data.frame that can be used to set width, height,
units, and/or asp.}
    \item{\code{class}}{Class of return object: "data.frame" (default) or "list" (only supported when  )}
  }}
}
\description{
Coerce a character or named vector to a page data.frame using get_page_size()
or make_page_size()
}
\examples{
as_page(c(8.5, 11), units = "in")

as_page("letter")

as_page(
  list(name = "letter", w = 8.5, h = 11, units = "in"),
  cols = c("w", "h")
)

}
