% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dist_scale.R
\name{convert_dist_scale}
\alias{convert_dist_scale}
\title{Convert distance from scale to actual units}
\usage{
convert_dist_scale(
  dist = NULL,
  scale = NULL,
  standard = NULL,
  series = NULL,
  scale_unit = "in",
  scale_factor = NULL,
  actual_unit = NULL,
  dpi = 120,
  paper = NULL,
  orientation = NULL,
  ...
)
}
\arguments{
\item{dist}{distance to convert. If paper is provided, dist is optional and
paper width and height are used as dist.}

\item{scale}{Scale name from \code{standard_scales[["scale"]]}.}

\item{standard}{Scale standard. Options include "USGS", "Engineering", or
"Architectural".}

\item{series}{Map series from \code{standard_scales[["series"]]}. Series is only
available for USGS scales.}

\item{scale_unit}{"mm" (converted to cm by dividing by 10), "cm", "px"
(converted to inches by dividing by dpi), or "in".}

\item{scale_factor}{factor for converting from scale_unit to actual_unit,
e.g. if 1" = 1', the scale factor is 12. optional if scale if provided;
defaults to \code{NULL}.}

\item{actual_unit}{any unit supported by \code{\link[=convert_dist_units]{convert_dist_units()}}}

\item{dpi}{dots per square inch (used as conversion factor for "px" to "in")}

\item{paper}{Name of paper passed to \code{\link[=get_paper]{get_paper()}}}

\item{orientation}{Page orientation, Default: NULL}

\item{...}{
  Arguments passed on to \code{\link[=get_paper]{get_paper}}
  \describe{
    \item{\code{name}}{Page name, e.g. "letter", not case sensitive, Default: NULL}
    \item{\code{width}}{Page width in "in", "px" or "mm" units. Default: NULL}
    \item{\code{height}}{Page height in "in", "px" or "mm" units. Default: NULL}
  }}
}
\value{
\itemize{
\item If paper is not provided, return a vector of dist values converted from
scale_unit to actual_unit based on scale_factor or information from
\link{standard_scales} data.
\item If paper is provided, return a data.frame with converted distances appends
as columns named actual_width and actual_height.
}
}
\description{
This function converts scale distances to actual units based on named
\link{standard_scales}.
}
\seealso{
Other dist: 
\code{\link{convert_dist_units}()},
\code{\link{is_same_units}()}
}
\concept{dist}
