% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inset_page_element.R
\name{inset_page_element}
\alias{inset_page_element}
\title{Create an inset with page size dimensions be added on top of the previous plot}
\usage{
inset_page_element(
  p,
  inset_page = NULL,
  left = NULL,
  bottom = NULL,
  right = NULL,
  top = NULL,
  align_to = "panel",
  on_top = TRUE,
  clip = TRUE,
  ignore_tag = FALSE,
  ...
)
}
\arguments{
\item{p}{A grob, ggplot, patchwork, formula, raster, or nativeRaster object
to add as an inset}

\item{inset_page}{Page size data.frame to use for inset, Default: \code{NULL}}

\item{left, bottom, right, top}{numerics or units giving the location of the
outer bounds. If given as numerics and inset_page is NULL, they will be
converted to npc units. All four are required if inset_page is\code{ NULL}. If
inset_page is provided, top \emph{or} bottom and left \emph{or} right must be
provided as the inset element is expected to be the width and height
defined by inset_page.}

\item{align_to}{Specifies what \code{left}, \code{bottom}, etc should be relative to.
Either \code{'panel'} (default), \code{'plot'}, or \code{'full'}.}

\item{on_top}{Logical. Should the inset be placed on top of the other plot or
below (but above the background)?}

\item{clip}{Logical. Should clipping be performed on the inset?}

\item{ignore_tag}{Logical. Should autotagging ignore the inset?}

\item{...}{
  Arguments passed on to \code{\link[=convert_unit_type]{convert_unit_type}}
  \describe{
    \item{\code{from}}{Unit to convert from. If \code{NULL} and x is not a units object,
convert to \code{to} units with a warning.}
    \item{\code{to}}{Unit to convert to. Passed to unitTo parameter of
\code{\link[grid:grid.convert]{grid::convertUnit()}}. If \code{NULL}, return x as is.}
    \item{\code{typeFrom}}{Passed to typeFrom parameter of \code{\link[grid:grid.convert]{grid::convertUnit()}}.
Defaults to "dimension".}
    \item{\code{valueOnly}}{Passed to valueOnly parameter of \code{\link[grid:grid.convert]{grid::convertUnit()}}.
Defaults to \code{FALSE}.}
    \item{\code{x}}{A numeric vector.

    For \code{is.unit}, any R object.}
  }}
}
\value{
A \code{inset_path} object
}
\description{
Create an inset with page size dimensions be added on top of the previous plot
}
\examples{
\dontrun{
if (interactive() & is_installed("ggplot2")) {
  library(ggplot2)
  p <- ggplot(mpg, aes(displ, fill = class)) +
    geom_bar()

  ggplot(mpg, aes(displ, hwy, colour = class)) +
    geom_point() +
    inset_page_element(
      p = p,
      inset_page = get_page_size("Poker card", orientation = "landscape"),
      left = unit(1, "in"),
      bottom = unit(1, "in")
    )
}
}
}
\seealso{
\code{\link[patchwork]{inset_element}}
}
