% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dist_units.R, R/get_dist_units.R
\name{is_same_units}
\alias{is_same_units}
\alias{is_dist_units}
\alias{get_dist_units}
\title{General utility functions for working with distance units objects}
\usage{
is_same_units(x, y = NULL)

is_dist_units(x)

get_dist_units(x, multiple = TRUE, quiet = FALSE)
}
\arguments{
\item{x, y}{objects to check}

\item{multiple}{If \code{TRUE} and x is a character vector with distance/area
units, \code{\link[=get_dist_units]{get_dist_units()}} may return multiple units. Passed to
\link[rlang:arg_match]{rlang::arg_match}.}

\item{quiet}{If \code{TRUE}, suppress warning messages.}

\item{null.ok}{If null.ok is \code{TRUE}, allow x to return a \code{NULL} value; if
\code{FALSE}, error on \code{NULL} values.}
}
\description{
\itemize{
\item \code{\link[=is_dist_units]{is_dist_units()}}: Is x a distance unit object?
\item \code{\link[=is_same_units]{is_same_units()}}: are x and y character strings that represent the same
\item \code{\link[=get_dist_units]{get_dist_units()}}: Get the distance units from x (if x is a sf or units
objects or a character string from \link{dist_unit_options})
\item \code{\link[=as_dist_units]{as_dist_units()}}: Convert x to units using \link[units:units]{units::as_units}
}
}
\seealso{
Other dist: 
\code{\link{convert_dist_scale}()},
\code{\link{convert_dist_units}()}
}
\concept{dist}
