% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist_units.R
\name{is_dist_units}
\alias{is_dist_units}
\alias{get_dist_units}
\alias{as_dist_units}
\alias{is_same_units}
\title{General utility functions for working with distance units objects}
\usage{
is_dist_units(x, arg = caller_arg(x))

get_dist_units(x, arg = caller_arg(x), call = parent.frame())

as_dist_units(x, units = NULL, arg = caller_arg(x), call = parent.frame())

is_same_units(x, y = NULL)
}
\arguments{
\item{x, y}{objects to check}

\item{arg}{Used internally and passed to \code{\link[rlang:arg_match]{rlang::arg_match()}} as error arg or
used by \code{\link[cli:cli_abort]{cli::cli_abort()}} to improve error messages.}

\item{call}{Passed to error_call for \code{\link[rlang:arg_match]{rlang::arg_match()}}}

\item{units}{Distance units to convert to. Must be one of dist_unit_options
or area_unit_options.}
}
\description{
\itemize{
\item \code{\link[=is_dist_units]{is_dist_units()}}: Is x a units object with a units attribute in
\code{dist_unit_options} or \code{area_unit_options}?
\item \code{\link[=get_dist_units]{get_dist_units()}}: Get the distance units from x (if x is a sf or units
objects or a character string from \link{dist_unit_options})
\item \code{\link[=as_dist_units]{as_dist_units()}}: Convert x to units using \link[units:units]{units::as_units}
\item \code{\link[=is_same_units]{is_same_units()}}: Do x and y have the same distance units attribute?
Names or symbols of valid distance units are allowed.
}
}
\examples{

mile <- units::set_units(1, "mi")

is_dist_units("mi")

is_dist_units(mile)

is_same_units(mile, "mile")

get_dist_units(mile)

as_dist_units(1, "mi")

as_dist_units(2, mile)

}
\seealso{
Other dist: 
\code{\link{convert_dist_scale}()},
\code{\link{convert_dist_units}()}
}
\concept{dist}
