% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_contact_sheets.R
\name{make_contact_sheets}
\alias{make_contact_sheets}
\title{Use \code{magick::image_ggplot()} to make contact sheets for images}
\usage{
make_contact_sheets(
  images,
  dims,
  image_fileext = NULL,
  tags = NULL,
  captions = "file_name",
  caption_size = 12,
  caption_position = "panel",
  image_margin = margins(0.1, unit = "in"),
  page = "letter",
  orientation = "portrait",
  save = FALSE,
  filename = NULL,
  ...
)
}
\arguments{
\item{images}{File path or data.frame from \code{\link[filenamr:read_exif]{filenamr::read_exif()}}}

\item{dims}{Image dimensions in same dimensions as page. Required.}

\item{image_fileext}{Passed to fileext parameter of \code{\link[filenamr:read_exif]{filenamr::read_exif()}},
Default: \code{NULL}}

\item{tags}{Optional list of EXIF tags to read from files. Must include GPS
tags if to create an \code{sf} object based on the resulting data.frame object.}

\item{captions}{Template for caption, passed to \code{\link[glue:glue]{glue::glue()}} for execution
within function environment. Default: '{images$file_name}
'{images$create_date}'}

\item{caption_size}{Caption size, passed to \code{\link[ggplot2:element]{ggplot2::element_text()}} for plot.caption
for theme, Default: 12}

\item{caption_position}{Caption position, passed to plot.caption.position for
theme, Default: 'panel'}

\item{image_margin}{Image margin passed Default: \code{margins(0.1, unit = "in")}}

\item{page}{Paper name or a data.frame with width and height columns.
Optional if width and height are both provided, Default: \code{NULL}}

\item{orientation}{Paper orientation, Optional if width and height are both
provided, Default: 'landscape'}

\item{save}{If \code{TRUE}, save contact sheet to a file. filename may be required
if save is \code{TRUE}. Default: \code{FALSE}}

\item{filename}{File name to create on disk.}

\item{...}{Additional parameters passed to \code{\link[=map_ggsave_ext]{map_ggsave_ext()}} excluding
width, height, and units.}
}
\value{
A list of patchwork object or (if save = TRUE) invisibly return the
list and save a file.
}
\description{
Wraps \code{\link[filenamr:read_exif]{filenamr::read_exif()}}, \code{\link[magick:image_ggplot]{magick::image_ggplot()}}, and \code{\link[=page_layout]{page_layout()}}
to create contact shets for a folder of images.
}
\seealso{
\code{\link[filenamr:read_exif]{filenamr::read_exif()}}
\code{\link[glue:glue]{glue::glue()}}
\code{\link[magick:editing]{magick::editing()}}, \code{\link[magick:image_ggplot]{magick::image_ggplot()}}
\code{\link[ggplot2:labs]{ggplot2::labs()}}, \code{\link[ggplot2:theme]{ggplot2::theme()}}, \code{\link[ggplot2:element]{ggplot2::margin()}}
}
