% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_layout.R
\name{page_layout}
\alias{page_layout}
\alias{layout_cards}
\title{Use patchwork to lay out a list of fixed aspect plots on a larger page}
\usage{
page_layout(
  plots = NULL,
  page = NULL,
  width = NULL,
  height = NULL,
  orientation = "landscape",
  byrow = FALSE,
  guides = NULL,
  tag_level = NULL,
  design = NULL,
  paginate = TRUE,
  dims = NULL,
  images = FALSE,
  dpi = 120,
  ...
)
}
\arguments{
\item{plots}{Page name, a data.frame with width and height columns, or a list
of ggplot2 objects with card plots. Default: \code{NULL}}

\item{page}{Paper name or a data.frame with width and height columns.
Optional if width and height are both provided, Default: \code{NULL}}

\item{width, height}{Paper width and height, Default: \code{NULL}}

\item{orientation}{Paper orientation, Optional if width and height are both
provided, Default: 'landscape'}

\item{byrow}{Analogous to \code{byrow} in \link[base:matrix]{matrix()}. If \code{FALSE} the
plots will be filled in in column-major order}

\item{guides}{A string specifying how guides should be treated in the layout.
\code{'collect'} will collect guides below to the given nesting level, removing
duplicates. \code{'keep'} will stop collection at this level and let guides be
placed alongside their plot. \code{auto} will allow guides to be collected if a
upper level tries, but place them alongside the plot if not.  If you modify
default guide "position" with \link[ggplot2:theme]{theme(legend.position=...)}
while also collecting guides you must apply that change to the overall
patchwork (see example).}

\item{tag_level}{A string (\code{'keep'} or \code{'new'}) to indicate how
auto-tagging should behave. See \code{\link[patchwork:plot_annotation]{plot_annotation()}}.}

\item{design}{Specification of the location of areas in the layout. Can either
be specified as a text string or by concatenating calls to \code{\link[patchwork:area]{area()}} together.
See the examples for further information on use.}

\item{paginate}{If \code{TRUE}, create a list of \code{patchwork} objects when the
number of plots is greater than the number of spaces in the plot layout.
Default to \code{TRUE}.}

\item{dims}{Optional. Plot dimensions. If \code{NULL} (default), dims are inferred
based on the dimensions of the first plot in plots.}

\item{images}{If \code{TRUE} and dims is \code{NULL}, the input plots are assumed to be
plots created with \code{\link[magick:image_ggplot]{magick::image_ggplot()}} and dpi is used to infer
dimensions.}

\item{dpi}{If images is \code{TRUE} and dims is \code{NULL},}

\item{...}{Additional parameters passed to \code{\link[=get_page_size]{get_page_size()}} if dims is a
character object.}
}
\value{
A \code{patchwork} object or a list of \code{patchwork} objects.
}
\description{
Use patchwork to lay out a list of fixed aspect plots on a larger page
}
\examples{
page_layout(
  plots = plot_cards("Poker", 6),
  page = "letter"
)
}
\seealso{
\code{\link[ggplot2:ggplot_build]{ggplot2::ggplot_build()}}
\code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}, \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}
}
