% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_page.R
\name{set_page_dims}
\alias{set_page_dims}
\alias{set_page_orientation}
\alias{set_page_asp}
\title{Set page data.frame dimensions, orientation, or aspect ratio}
\usage{
set_page_dims(
  page,
  dims = NULL,
  width = NULL,
  height = NULL,
  units = NULL,
  cols = c("width", "height")
)

set_page_orientation(
  page,
  orientation = NULL,
  tolerance = 0.1,
  cols = c("width", "height")
)

set_page_asp(page, flipped = FALSE, cols = c("width", "height"))
}
\description{
Set page data.frame dimensions, orientation, or aspect ratio

set_page_asp: appends an aspect ratio column a page size data.frame
}
