% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_page_margins.R
\name{set_page_margin}
\alias{set_page_margin}
\title{Set margins for page data.frame (adding body width, height, and asp)}
\usage{
set_page_margin(
  page = NULL,
  margins,
  unit = "in",
  cols = c("width", "height"),
  ...
)
}
\arguments{
\item{page}{A character vector with a page size name or a data.frame. Passed
to x parameter of \code{\link[=as_page]{as_page()}}.}

\item{margins}{Passed to \code{\link[=get_margin]{get_margin()}} with unit value.}

\item{unit}{Unit used for the margin. If margin is a unit object, unit is
ignored. If page uses different units, the margins are converted into the
page units for consistency.}

\item{cols}{Column names to use for width and height columns. Defaults to
c("width", "height"). Must be length 2 and the first value is always used
as as the width name and the second as the height.}

\item{...}{Passed to \code{\link[=as_page]{as_page()}} with page and cols.}
}
\value{
A data.frame with the page dimensions and additional columns for
body dimensions, body aspect ratio, and margins.
}
\description{
Set margins for page data.frame (adding body width, height, and asp)
}
\seealso{
\code{\link[ggplot2:element]{ggplot2::margin()}}; \code{\link[=set_page_dims]{set_page_dims()}}; \code{\link[=set_page_orientation]{set_page_orientation()}};
\code{\link[=set_page_asp]{set_page_asp()}}
}
