% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_page_margins.R
\name{set_page_margin}
\alias{set_page_margin}
\title{Set margins for page paper with margin}
\usage{
set_page_margin(
  page = NULL,
  margins = NULL,
  unit = "in",
  cols = c("width", "height"),
  ...
)
}
\arguments{
\item{margins}{Passed to \code{\link[=get_margin]{get_margin()}} with unit}
}
\description{
Set margins for page paper with margin
}
